Move most bins reduction functions into the variable module
