+++ lib/cmake/scipp-functions.cmake
@@ -86,28 +86,28 @@ scipp_function(
   "unary" bins bin_sizes SKIP_VARIABLE BASE_INCLUDE variable/bins.h
 )
 scipp_function(
-  "unary" bins bins_mean SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_mean SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 scipp_function(
-  "unary" bins bins_sum SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_sum SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 scipp_function(
-  "unary" bins bins_max SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_max SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 scipp_function(
-  "unary" bins bins_nanmax SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_nanmax SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 scipp_function(
-  "unary" bins bins_min SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_min SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 scipp_function(
-  "unary" bins bins_nanmin SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_nanmin SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 scipp_function(
-  "unary" bins bins_all SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_all SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 scipp_function(
-  "unary" bins bins_any SKIP_VARIABLE BASE_INCLUDE dataset/bins_reduction.h
+  "unary" bins bins_any SKIP_VARIABLE BASE_INCLUDE variable/reduction.h
 )
 setup_scipp_category(bins)
